Setting up mutation requests
