+++ KsApi/ServiceType.swift
@@ -42,6 +42,9 @@ public protocol ServiceType {
   func addVideo(file fileURL: URL, toDraft draft: UpdateDraft)
     -> SignalProducer<UpdateDraft.Video, ErrorEnvelope>
 
+  func changePassword(input: GraphMutationInput) ->
+    SignalProducer<GraphMutationEmptyResponseEnvelope, GraphError>
+
   func changePaymentMethod(project: Project)
     -> SignalProducer<ChangePaymentMethodEnvelope, ErrorEnvelope>
 
@@ -400,6 +403,40 @@ extension ServiceType {
       return self.preparedRequest(forRequest: request, queryString: queryString)
   }
 
+  /**
+   Prepares a URL request to be sent to the server.
+   - parameter originalRequest: The request that should be prepared
+   - parameter queryString: The GraphQL mutation string description
+   - parameter input: The input for the mutation
+
+   - returns: A new URL request that is properly configured for the server
+ **/
+  public func preparedGraphRequest(forURL url: URL,
+                                   queryString: String,
+                                   input: [String: Any]) throws -> URLRequest {
+    var request = URLRequest(url: url)
+    request.httpMethod = Method.POST.rawValue
+
+    guard let URL = request.url else {
+      return request
+    }
+
+    let requestBody = self.graphMutationRequestBody(mutation: queryString, input: input)
+    let jsonData = try JSONSerialization.data(withJSONObject: requestBody, options: [])
+
+    request.httpBody = jsonData
+
+    var headers = self.defaultHeaders
+    headers["Content-Type"] = "application/json; charset=utf-8"
+
+    // swiftlint:disable:next force_unwrapping
+    let components = URLComponents(url: URL, resolvingAgainstBaseURL: false)!
+    request.url = components.url
+    request.allHTTPHeaderFields = headers
+
+    return request
+  }
+
   public func isPrepared(request: URLRequest) -> Bool {
     return request.value(forHTTPHeaderField: "Authorization") == authorizationHeader
       && request.value(forHTTPHeaderField: "Kickstarter-iOS-App") != nil
@@ -414,6 +451,13 @@ extension ServiceType {
     headers["User-Agent"] = Self.userAgent
 
     return headers
+  }
+
+  fileprivate func graphMutationRequestBody(mutation: String, input: [String: Any]) -> [String: Any] {
+    return [
+    "query": mutation,
+    "variables": ["input": input]
+    ]
   }
 
   public static var userAgent: String {
